Added MatrixGenerator Class, added speedtest to main
I seperated the main into two different tests, one is for speed, one is for
quality. The matrixGenerator class can generate random matrix files
or produce DataObj class objects directly. Since we are using doubles
the heap size of java is pretty limited, so we can only test with a
5000 x 5000 matrix. I didn't want to deal with reloading data from disc
while the algorithm is running and changing from double to integer
seems also a bit too much work now. The time values should be meaningful
enough like this anyway.